Major Changes
I added a very simple player health system for you to use. Just reference the public Take damage function in the script.

I removed Player Input script because it was hard to understand and rewrote the camera and player script

I added jumping but it's not smooth at all

I added a camera transition when the player is in a combat state using triggers. it's in the Room Trigger script