Programming 2/2
Move Ui art and code to a separate and removed unused code

In Progress :
Camera Collision  works great when it comes to exploring, the clamps to a distance of 4 units away for the player after dividing 8 by 2. Works the same for Combat but we clamp it to 7.5. It's still needs to be tested and revised for all possible outcomes

Started working on Item wheel

Attacking Lunge: basically functional for quick code, the player teleports to it's new position about 5 unit up, will created a smooth transition later